Move category mutations
