--- saleor/dashboard/graphql/product/forms.py
@@ -1,9 +1,0 @@
-from django import forms
-
-from ....product.models import Product
-
-
-class ProductForm(forms.ModelForm):
-    class Meta:
-        model = Product
-        exclude = ['updated_at']
