+++ tests/dashboard/test_graphql.py
@@ -57,136 +57,6 @@ def test_model_form_update_mutation(model_form_class):
     assert 'id' in meta.arguments
 
 
-def test_category_create_mutation(admin_client):
-    query = """
-        mutation($name: String!, $description: String, $parentId: ID) {
-            categoryCreate(
-                name: $name
-                description: $description
-                parentId: $parentId
-            ) {
-                category {
-                    id
-                    name
-                    slug
-                    description
-                    parent {
-                        name
-                        id
-                    }
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-
-    category_name = 'Test category'
-    category_description = 'Test description'
-
-    # test creating root category
-    variables = json.dumps({
-        'name': category_name, 'description': category_description})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryCreate']
-    assert data['errors'] == []
-    assert data['category']['name'] == category_name
-    assert data['category']['description'] == category_description
-    assert not data['category']['parent']
-
-    # test creating subcategory
-    parent_id = data['category']['id']
-    variables = json.dumps({
-        'name': category_name, 'description': category_description,
-        'parentId': parent_id})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryCreate']
-    assert data['errors'] == []
-    assert data['category']['parent']['id'] == parent_id
-
-
-def test_category_update_mutation(admin_client, default_category):
-    query = """
-        mutation($id: ID, $name: String!, $description: String) {
-            categoryUpdate(
-                id: $id
-                name: $name
-                description: $description
-            ) {
-                category {
-                    id
-                    name
-                    description
-                    parent {
-                        id
-                    }
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-    # create child category and test that the update mutation won't change
-    # it's parent
-    child_category = default_category.children.create(name='child')
-
-    category_name = 'Updated name'
-    category_description = 'Updated description'
-
-    category_id = graphene.Node.to_global_id('Category', child_category.pk)
-    variables = json.dumps({
-        'name': category_name, 'description': category_description,
-        'id': category_id})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryUpdate']
-    assert data['errors'] == []
-    assert data['category']['id'] == category_id
-    assert data['category']['name'] == category_name
-    assert data['category']['description'] == category_description
-
-    parent_id = graphene.Node.to_global_id('Category', default_category.pk)
-    assert data['category']['parent']['id'] == parent_id
-
-
-def test_category_delete_mutation(admin_client, default_category):
-    query = """
-        mutation($id: ID!) {
-            categoryDelete(id: $id) {
-                category {
-                    name
-                }
-                errors {
-                    field
-                    message
-                }
-            }
-        }
-    """
-    variables = json.dumps({
-        'id': graphene.Node.to_global_id('Category', default_category.id)})
-    response = admin_client.post(
-        reverse('dashboard:api'), {'query': query, 'variables': variables})
-    content = get_graphql_content(response)
-    assert 'errors' not in content
-    data = content['data']['categoryDelete']
-    assert data['category']['name'] == default_category.name
-    with pytest.raises(default_category._meta.model.DoesNotExist):
-        default_category.refresh_from_db()
-
-
 def test_page_create_mutation(admin_client):
     query = """
         mutation CreatePage(
